ExcessiveLogonFailures - change Computer/IpAddress from maklist to make_set, so only unique values are logged
makelist provides upto 128 values, but its not deduped.
make_set provides deduped values, and I set the limit to 128 to match makelist. If desired this number can be lowered but I kept it to match makelist. In general, make_set with a limit of 128 will provide many more unique values.