Update Comment-MCASAlertURL
Removing reference to the required MCAS connection. The required connection is for Sentinel and Azure Monitor Logs. MCAS will be needed as a data source to feed the right logs into Azure Monitor to get value from the logic app but does not impact the logic app from running successfully.