Change the blob name to allow easier sorting
I suggest to do a zero-padding of the hour part of the blog name to allow better sorting. Example of the current output generates the following blob name pattern: 
   WindowsFirewall-2020-05-21-2.json
When we order blobs by name, the WindowsFirewall-2020-05-21-10.json will be listed before the WindowsFirewall-2020-05-21-2.json. The change will make it:
   WindowsFirewall-2020-05-21-02.json
This will allow proper lexicographic sorting.
The change consists of replacing the section with the HoursCount variable by the following expression:
if( less( variables('HoursCount'), 10), concat(variables('HoursCount'),'0'),variables('HoursCount'))
Note that this suggested notation will also reflect the recommendations from the ISO 8601 standard.