Update AnomalousIPUsageFollowedByTeamsAction.kql
Refactored code substaintially to remove bug in code, implemented feedback from Shain along the way.

Edited delta threshold default to 95%.

Edited execution window, works better with more data so 30 days by default.

Fluctuating results bug:
I was making a JSON object using make_bag which was using the number of times an IP was observed as a key (this is because I used this later to find the lowest hitting IP). When a high prevalence country IP and a low prevalence country IP both had the same number of authentications this led to the make_bag function randomly picking one. e.g. Sometimes and account had 2 logins from a ZA IP and sometimes from a GB IP - when the GB one was picked, it threw the result away.