SimpleMemberName should account for case sensitivity
The definition of SimpleMemberName in its present form fails to account for distinguished names containing either the member name or the string 'CN' in lower case.

extend SimpleMemberName = tostring(split(tostring(split(MemberName, ",")[0]),"CN=")[1])

For instance - event ID 4728 was logged in my lab for adding a user to Domain Admins group. However, the string 'CN' was in lower case in the distinguished name for the member name. As a result, the columns like SimpleMemberName, AccountCustomEntity were empty in the output corresponding to this event.