Change the location used to create the data connection
The current version is using the location of the log analytic workspace. That works only if the ADX cluster is also in the same location.
For this to work in cases where the ADX cluster is in a different location as the log analytic workspace we need to provide the ADX's location.
I suggest to extract the location from $ADXClusterURL and provide it in the location parameter when swending the PUT to create the data connection.