Update RDP_MultipleConnectionsFromSingleSystem.yaml
The original query misses out on scenarios where there were 0 RDP connections in the previous 7 days from the same account and IP address. This change takes that into account and behaves as if there had been 1 such RDP connection in the previous 7 days. This fits with the original query because the original query would produce a result if there are 2 (or more) RDP connections today with only 1 RDP connection in the previous 7 days.