Update IPEntity_AzureSQL.yaml
Removed account for now, not sure there is a mapping for account, Ajeet to look at when he is back.  Fixed indent on entitymapping.  Moved | where TimeGenerated >= ago(dt_lookBack) up in the AzureDiagnostic query as time is the primary index. Specifying join kind explicitly and commenting as to why we use innerunique. Making changes to filtering after the join and fixing the 2nd argmax to use the activity time instead of indicator time as  you want the most recent activity that matches, both changes per discussions on another PR #3477.