Update ChiaCryptoMining.yaml
Based on issue #3455: Chia rule entity mapping issues - Fixing entity mappings, changing process to file as we do not have a process name entity mapping, adding in FilePath and we are missing a column value to indicate the hash algorithm so added that in.  UX will look for xxxCustomEntity to auto map some values as part of the entity mapping feature so making sure those are available still. Also reducing extending names so it is done once outside of the join, this also reduces characters below 10000 which the rule creator errors on if more than.