Updated with latest logo image.
Updated with latest logo image.

Also, just an FYI - We had a meeting with Eric Burkholder yesterday about having our contribution as a Solutions or leave as is. And it was decided that we will have it as is (not a Solution) for now. And then move it at a later date.