Revoked IOCs are not excluded.
Lines 32 and 33 must be in the reverse order.  

If we check whether an IOC is revoke before summarizing the latest record we will always get the older version of the IOC after it has been revoked. Changing the order of these two lines ensures that we have the latest version of the IOC before we check if that version is active.