Fixes for ASimAuthenticationMicrosoftWindowsEvent
Various items fixed:

WindowsEvent contains EventOriginId as a field, hence the project-rename for EventID to EventOriginId was failing, resulting in no records being found at all. Fixed by removing this from project-rename and uncommented rename for EventOriginId to EventOriginalUid which seemed to be in the kusto query already.  Also put EventID in EventOriginID + dropped EventID as per consistency for how SecurityEvents are handled.

ActorUsername and TargetUsername were "extended" twice. Moreover, 2nd extend of TargetUsername was relying on EventData.TargetUsername which is faulty as it's EventData.TargetUserName (upper N) resulting in empty Targetusername. Fixed by removing the first 2 extends for these fields and correctly implementing lower in the query.

TargetDomainName was blank in my forwarded events, resulting in Windows to be used for TargetUsernameType rather than Simple. Fixed this by not only looking for a "-" but also "" (blank string)