Create codeql-analysis.yml
Code scanning is a feature that we can use to analyze the code in a GitHub repository to find security vulnerabilities and coding errors. Any problems identified by the analysis are shown in GitHub. We can use code scanning to find, triage, and prioritize fixes for existing problems in the code. Code scanning also prevents developers from introducing new problems. We can schedule scans for specific days and times, or trigger scans when a specific event occurs in the repository, such as a push.