Update AuthenticationMethodsChangedforPrivilegedAccount.yaml
Target entities were not being mapped due to aggregation. Provided an alternative with comments, using the same output schema (Except Targets changes between dynamic and string depending on which line is used). This allows for Target users to be mapped to entities if required.

Also added entity mapping for the Initiator account