Update azuredeploy_Connector_Proofpoint_AzureFunction.json
Increasing maxLength of the function app. We are facing issue with some client that they have some specific naming convention due to function app length size they are not able to deploy using template because there naming convention of function app having more than 11 char. So need to update the maxLeanght size of the function app.