Fix ProofpointPOD_message_CL parsing
"isnotempty( pps_cid_s)" doesn't work in this case because the ProofpointPOD_message_CL table doesn't have a pps_cid_s column. For us, this meant the Proofpoint_POD_message_CL table isn't parsed and none of the analytic rules in the Proofpoint PoD solution could be enabled because is couldn't find the NetworkDirection.