Update Massdownload_USBFileCopy.yaml
Adjusting a bit so we only match if the USB drive has the files and cleaning up a bit of the project statements, then packing so we can list that info in a single field.  Removing Process entity as Filename isn't ProcessId and Command line impacts query performance and isn't specifically relevant in the result here.