Build 2.0.0 version of Tanium solution
This is still what was going into the 1.0.10 version but now using the
V2 build process. While I was updating all of the metadata for V2 I
figured this solution build is different enough from 1.0.9 that we
should bump a major version. This is the solution build that no longer
requires the forwarder VM.