MerakiParser - Wrong handling of EventEndTime epoch with nanoseconds
EventEndTime seems to be a bogus, as it is parsed from the part of Epoch field after the dot. From the examples from Cisco, this seems to be Epoch timestamp with nanoseconds after the decimal, and not an "endtime". https://documentation.meraki.com/General_Administration/Monitoring_and_Reporting/Syslog_Event_Types_and_Log_Samples

Graylog seems to drop events where Meraki is configured to log timestamps with nanoseconds (https://docs.graylog.org/docs/cisco-meraki), but in Sentinel's case it would be possible to discard the part after the dot, as the split is already in place.