New Playbook - Update-CVE-IPs-WatchListwithGreyNoise
This playbook uses the GreyNoise API to search for intresting IPs discovered in the last day tagged per each CVE found in the mode you setup. See below for more information on the Modes to deploy with. This is a daily timer  trigger to go out and fetch things and push to a watchlist, also comes with a sample detection analytic rule.