ActivityStatusValue appears to be changed/updated
Was having issues testing this rule in an environment and found that the `ActivityStatusValue` from table AzureActivity regarding workspaces/computes/delete appears to be formatted as `Success` instead of `Succeeded`

Not sure if we should evaluate for both messages (or if there are others) or is should be a case-insensitive match on partial like `Succe` to match both. 

The above proposed solution matches on both and is case-insensitive.