Update Dev-0530AVHits.yaml for performance
Flipped tables so that SecurityAlert is $left and DeviceInfo is $right for performance

The DeviceInfo table is significantly larger than SecurityAlert in most organizations, and a large join using DeviceInfo as $left was causing Log Analytics lookup failure leading to Auto Disablement of the analytic query. Per https://learn.microsoft.com/en-us/azure/data-explorer/kusto/query/lookupoperator, `join` assumes that $left is smaller than $right, which was not true in the previous commit. Alternatively, lookup assumes $left is the larger table.