modify guid for uniqueness
If customers use the template as is with an ID that is a GUID only based on subscriptionID, any new CCP will overwrite previous CCP made with this template, including the GitHub data connector. By revising the GUID to be based on resourcegroup and deployment name, additional use of this template for new CCP won't overwrite previous.