Update AMAmigrationTracker.json
The query that we have in line 367 is incorrect. The join argument is happening on AzureVMs only which results in all HybridCompute resources to be listed in the results regardless if they have 2 agents or not. I cant think the logic behind the "join" to be honest but the proposed changes make the query simpler to understand and give expected results for ARC machines:

resources
| where type has 'microsoft.compute/virtualmachines/extensions' or  type has 'microsoft.hybridcompute/machines/extensions'
| where name has 'MicrosoftMonitoringAgent' or name has 'OmsAgentForLinux' or name has 'AzureMonitorWindowsAgent' or name has 'AzureMonitorLinuxAgent'
| extend AzureVM = extract('virtualMachines/(.*)/extensions',1,id), ArcVM = extract('machines/(.*)/extensions',1,id)
| summarize count() by AzureVM=tolower(AzureVM), ArcVM=tolower(ArcVM), subscriptionId, resourceGroup 
| extend hasBoth = iff(count_ > 1, 'Yes', 'No')
| where hasBoth == "Yes"
| summarize count()