Fix FQDN aggregation for Port Sweep Firewall rule
This is similar to PR 8764, and this will undo a change from a recent
commit (a7132b0) and restore the "Port Sweep" Firewall analytics rule
FQDN aggregation. Currently the rule summarizes by the FQDN, which
results in the query returning zero results unless the
"MinimumDifferentHostsThreshold" is set to 0, thereby making the
parameter useless.