Update azuredeploy_Cloudflare_API_FunctionApp.json
Property for MAX_CHUNK_SIZE_MB which allows the variable to be set. The default of 2 in the function app code is too small. Cloudflare logs frequently come in files larger than that and the function app will hang up and reprocess the same logs over and over if it doesn't download the entire file.