Update LoginfromUsersfromDifferentCountrieswithin3hours.yaml
When running the query there were issues identified with how the artifacts were parsing out (outside the 3-hour time designation). The root issues were from the way the query was trying to parse using the timeframe string. The change removed the string | where column_ifexists('published_t', now()) >= timeframe and replaced it with | where TimeGenerated >= timeframe which correctly displayed the artifacts from the 3-hour time designation.

Improvements to the are showing more context to the alert such as the countries and the Lat and Long.