Update ForescoutEvent.yaml
I haven't had much success understanding what ```ACTIONidentity``` was supposed to be resolving to and so since Forescout was something that I could identify somewhat easily in the syslog table I chose to go down this route. Please review and let me know if this doesn't make sense. Appreciate it, thank you!