Update run.ps1 - fix LoginURL check
In azuredeploy.json the loginEndpoint dynamic parameter is set to `[environment().authentication.loginEndpoint]` which will return with a trailing `/` ([ref](https://learn.microsoft.com/en-us/azure/azure-resource-manager/bicep/bicep-functions-deployment)) and as such this test will always fail and halt the process.

Removing the "/" from the variable broke the auth. process. 

Suspect an issue was introduced in the pull from 8 months ago? #8933 

Will be necessary to re-create the zip as well, if this fix is accepted.