fix(workbooks) Updated limit on comply visualization to limit to top 10
Our visualization ```Top mission critical computers with CVSS v3 hits >= 9.0``` we had a limit of the top 4 results. I had removed this in a previous commit suspecting it was not intentional and to help me investigate what the raw data looks like. It has been decided will will limit to the top 10 but also update the visualization name to make it clear it's a limited list.