fix(workbook) Updated comply visualization to only use data from the latest run
Our visualation for `Vulnerability Distribution by Year`  is actually an update each time we get data. So for this one, we don't want to query any time range, we want to only see rows that were sent over in the last run. So I corrected this to find the date/time of the last run and limit the query that way.