Update AnomalousUserAppSigninLocationIncrease-detection.yaml
Provided a let variable to adjust the dcount accuracy argument with ease. 

Also set the "| where TimeGenerated" to use a between operator. When looking at the events from incidents within Sentinel (By selecting "Link to LA"), the query is evaluating logs from the 7 days preceding the original analytic rule run datetime to the current day, resulting in inaccurate representation of the events at that given time.