build(solution) Updated build script to include v3 script and addl checks
We had an old build script and another to check our build. The check script is deprecated due to the v3 build scripts from MSFT. So I remove it. Then I updated our build script to use the v3 build and to also complete a few more checks for us to help minimize human error.

# Conflicts:
#	Solutions/Tanium/check_build.sh