Tanium 3.1.0 Rebuild: Azure Team Breaking Workbook Images
With the changes from the team managing the repo two entries ended up in the WorkbooksMetadata.json file for workbook preview images and the Partner Center wouldn't accept our package zip. So we tried submitting a new build and removed the duplicates. Additionally, the team handled the reported issue of Workbook preview images not displaying. This fix resulted in using a different property name.  However, the team didn't update their validation scripts to match the change and our PR became blocked. They have now changed how preview images work again, and now the workbook previews don't show. But we have been waiting weeks to get this release out, so this rebuild is to try to get the playbook changes we made released, and handle the workbook images once that functionality and validation is fixed on Azure's side.